requirements.txt
flask==2.0.2
jieba==0.42.1
numpy==1.19.2
transformers==4.12.5
bert4keras==0.10.8
pandas==1.1.5
six==1.16.0
pyyaml==6.0
rouge==1.0.1
tqdm==4.62.3
networkx==2.5.1
torchsummary==1.5.1
textrank4zh==0.3